Changed the representation of the price to BigDecimal. Adjusted the selection flow to allow for a selection of cuisine before showing any item. Improved the flow of the user interface.

+++ src/main/java/exercises/xf/model/Dessert.java
@@ -1,11 +1,13 @@
 package exercises.xf.model;
 
+import java.math.BigDecimal;
+
 /**
  * Created by guisil on 31/01/2017.
  */
 public class Dessert extends Supply {
 
-    Dessert(String name, double price, Cuisine cuisine) {
+    Dessert(String name, BigDecimal price, Cuisine cuisine) {
         super(name, price, cuisine);
 
         if (cuisine == null) {
