+++ src/main/java/exercises/xf/model/PricedItem.java
@@ -1,10 +1,12 @@
 package exercises.xf.model;
 
+import java.math.BigDecimal;
+
 /**
  * Created by guisil on 31/01/2017.
  */
 public interface PricedItem {
 
     String getName();
-    double getPrice();
+    BigDecimal getPrice();
 }
